+++ src/api/type-defs/inputs/SetTimestampOffsetBatchTaskInput.ts
@@ -1,6 +1,6 @@
 export default /* GraphQL */ `
   input SetTimestampOffsetBatchTaskInput {
     imageIds: [ID!]!
-    offsetMs: Int!
+    offsetMs: Float!
   }
 `;
